==> notebooks/1) IMDB_TMDB Data Processing.ipynb <==
## IMDB TMDB Movie Dataset Data Processing

### Import Libraries

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
import plotly
import plotly.express as px
import plotly.graph_objects as go
import tensorflow as tf
from tensorflow import keras
from vaderSentiment.vaderSentiment import SentimentIntensityAnalyzer
import ast
import datetime
import sklearn
import scipy
import warnings
from tqdm.notebook import tqdm
warnings.filterwarnings('ignore')
tqdm.pandas()

%%html
<style>
.cell-output-ipywidget-background {
    background-color: transparent !important;
}
:root {
    --jp-widgets-color: var(--vscode-editor-foreground);
    --jp-widgets-font-size: var(--vscode-editor-font-size);
}  

df  = pd.DataFrame(pd.read_csv('../data/IMDB TMDB Movie Metadata Big Dataset (1M).csv'))
df.head()

df.shape

df.iloc[:,:20].head()

df.iloc[:,20:40].head()

df.iloc[:,40:].head()

df.dtypes

### Type Conversion

type(df['release_date'][0])

`release_date` is in `str` format. Let's change that to its original format.

df['release_date'] = pd.to_datetime(df['release_date'], errors='coerce')
print(type(df['release_date'][0]))
print(df['release_date'].head())

array_df = df[['keywords', 'genres_list', 'Cast_list', 'all_combined_keywords']]
array_df.head()

array_df.iloc[0].progress_apply(type)

The `array` are in `str` format. So we will need to change them back.

array_df = array_df.applymap(ast.literal_eval)
array_df.head()

for attribute in tqdm(list(array_df.keys())):
    df[attribute] = array_df[attribute]

print(array_df.iloc[0].apply(type))
df[['keywords', 'genres_list', 'Cast_list', 'all_combined_keywords']].head()

### Data Cleaning

df['status'].value_counts()

We only care about movies than have been released and generated revenue to measure success with.

df = df[df['status'] == 'Released']
df.shape

df.isnull().sum()

IMDB Dataset Attribute Analysis

Potentially Unnecessary Attributes

1. AverageRating (1,060,504 nulls)
- **Explanation:** This attribute is almost entirely null. It's redundant with `vote_average` and `IMDB_Rating`, which have no null values.

2. Poster_Link (1,069,485 nulls)
- **Explanation:** Nearly all entries are null. The `poster_path` attribute likely serves a similar purpose with fewer null values.

3. Certificate (1,042,420 nulls)
- **Explanation:** Mostly null. If certification information is crucial, it might be better to source it from a more complete dataset.

4. IMDB_Rating (1,041,943 nulls)
- **Explanation:** Largely null and likely redundant with `vote_average`.

5. Meta_score (1,069,924 nulls)
- **Explanation:** Almost entirely null. If Metacritic scores are desired, a more complete source should be used.

6. Music_Composer (916,779 nulls)
- **Explanation:** Largely null. This information might be included in the `Producers` or `Cast_list` attributes.

7. Director_of_Photography (735,402 nulls)
- **Explanation:** Mostly null. This level of detail might not be necessary for general analysis.

8. homepage (956,736 nulls)
- **Explanation:** Mostly null and not crucial for most analyses.

9. tagline (919,472 nulls)
- **Explanation:** Largely null and often not essential for analysis. 

8. Producers (646,947 nulls)
- **Explanation:** Mostly null.

9. Writer (425,675 nulls)
- **Explanation:** While it has fewer nulls than some other attributes, it's still mostly empty and might be redundant with information in Cast_list or overview.

<br>

Attributes to Consider Keeping Despite High Null Counts

1. production_countries (456,376 nulls)
- While it has many nulls, this information could be valuable for regional analysis.

2. spoken_languages (440,257 nulls)
- This could be important for language-based analysis, despite the high null count.

3. imdb_id (482,891 nulls)
- This could be useful for linking to external IMDB data, even though it has many nulls.

4. backdrop_path (779,429 nulls)
- While mostly null, it might be useful for visual representations in some applications.

5. release_date and release_year (both 150,556 nulls)
- These are important for temporal analysis, and the null count is relatively low compared to the dataset size.

6. Star1, Star2, Star3, Star4 (all have 1,041,942 or more nulls)
- **Explanation:** These are mostly null and redundant with the `Cast_list` attribute, which has no null values. But are important in determining the potential popularity of a movie based on its star-studded cast.

# Assuming your DataFrame is named 'imdb_df'

unnecessary_attributes = [
    'AverageRating',
    'Poster_Link',
    'Certificate',
    'IMDB_Rating',
    'Meta_score',
    'Music_Composer',
    'Director_of_Photography',
    'homepage',
    'tagline',
    'Producers',
    'Writer'
]

# Drop the unnecessary columns
df2 = df.drop(columns=unnecessary_attributes)
df2.columns

df2.isnull().sum()

df[['release_date', 'release_year']].head()

df[['release_date', 'release_year']].tail()

df[['release_date', 'release_year']].isnull().sum()

df[(df['release_date'].isnull()) & (df['release_year'].isnull())].isnull().sum()

The time data of this dataset is important to us, and thus we will not be able to use those rows without a date.

df2.dropna(subset=['release_date'], inplace=True)
df2.dropna(subset=['release_year'], inplace=True)
df2.shape

df2[['Star1', 'Star2', 'Star3', 'Star4']].head()

df2[['Star1', 'Star2', 'Star3', 'Star4']].tail()

This identifies if there are any popular/famous movie starts starring in the movie. So it is ok to have `null` values.

df2['Star1'] = df2['Star1'].fillna('Unknown')
df2['Star2'] = df2['Star2'].fillna('Unknown')
df2['Star3'] = df2['Star3'].fillna('Unknown')
df2['Star4'] = df2['Star4'].fillna('Unknown')
df2[['Star1', 'Star2', 'Star3', 'Star4']].tail()

df2[['production_countries', 'production_companies']].head()

df2[['production_countries', 'production_companies']].tail()

df2[(df2['production_companies'].isnull()) & (df2['production_countries'].isnull())].isnull().sum()

df2['production_companies'] = df2['production_companies'].fillna('Unknown')
df2['production_countries'] = df2['production_countries'].fillna('Unknown')
df2[['production_companies', 'production_countries']].tail()

df2.isnull().sum()

df2[['backdrop_path', 'poster_path']].head()

df2[['backdrop_path', 'poster_path']].tail()

On more intricate analysis, these attributes do not contribute to the project.

df2 = df2.drop(columns=['backdrop_path', 'poster_path'])
df2.columns

df2[['imdb_id', 'overview', 'spoken_languages']].head()

df2[['imdb_id', 'overview', 'spoken_languages']].tail()

`imdb_id` is not useful for further research, while `overview` and `spoken_languages` are still useful.

df2 = df2.drop(columns=['imdb_id'])
df2['overview'] = df2['overview'].fillna('Unknown')
df2['spoken_languages'] = df2['spoken_languages'].fillna('Unknown')
df2.isnull().sum()

df2 = df2.drop_duplicates(subset='title')
df2.shape

df2['actor_list'] = df2.progress_apply(lambda x: [x['Star1'], x['Star2'], x['Star3'], x['Star4']] + x['Cast_list'], axis=1)
print(df2['actor_list'][0])

df2 = df2[df2['adult'] == False]
df2 = df2[df2['status'] == 'Released']
df2.shape

df2['director'] = df2['Director'].progress_apply(lambda x: x.split(', '))
df2['genre'] = df2['genres_list']

df2.columns

df2['production_companies'] = df2['production_companies'].progress_apply(lambda x: x.split(', '))
df2['production_countries'] = df2['production_countries'].progress_apply(lambda x: x.split(', '))

df2.drop(columns=['overview', 'original_title', 'all_combined_keywords', 'adult', 'status', 'release_year', 'spoken_languages', 'Star1', 'Star2', 'Star3', 'Star4', 'Director', 'Cast_list', 'genres_list'], inplace=True)
df2.head()

df2_1 = df2.copy()

### Relevant Movies Filtering

We will also further limit our dataset to contain the movies that have been voted for more than 100 times, just so that we know that they are relevant.

print(df2.shape)
df2 = df2[df2['vote_count'] >= 100]
print(df2.shape)

We also want the movies to be relevant to this day and age. Thus, we will keep the movies from 2000 onwards.

print(df2.shape)
df2 = df2[df2['release_date'] >= '2000-01-01']
print(df2.shape)

print(df2_1.shape)
df2_1 = df2_1[df2_1['release_date'] >= '2000-01-01']
print(df2_1.shape)

df3 = df2
df3.columns

### Row Expansion

df3.shape

for i in tqdm(df3.columns):
    print(f'{i}: {type(df3[i][0])}, {df3[i][0]}')

df3 = df3.explode('production_companies').explode('production_countries').explode('keywords').explode('actor_list').explode('director').explode('genre')
df3.shape

df3.head()

# Identify the first occurrence in each group
first_occurrence = df3.groupby('id').cumcount() == 0

# Use the mask to set numerical attributes to NaN where it's not the first occurrence
numerical_attributes = df3.select_dtypes(include=[np.number]).columns
df3.loc[~first_occurrence, numerical_attributes] = np.nan

print("Transformed DataFrame:")
df3.head()

### Export Data

df2.to_csv('../data/imdb_tmdb_cleaned.csv', index=False)

df3.to_csv('../data/imdb_tmdb_tableau.csv', index=False)

df2_1.to_csv('../data/imdb_tmdb_cleaned_w|o_filtering.csv', index=False)

# # Open the file in write mode
# with open('../data/imdb_tmdb_cleaned.csv', 'w', newline='', encoding='utf-8') as f:
#     # Write the header first
#     df3.iloc[:0].to_csv(f, index=False)
    
#     # Use tqdm to track the progress of writing rows to CSV
#     for i in tqdm(range(0, len(df3))):
#         df3.iloc[[i]].to_csv(f, header=False, index=False)
