==> notebooks/2) Rotten Tomatoes Critic Reviews Data Processing.ipynb <==
## Rotten Tomatoes Critic Reviews Data Processing

### Import Libraries

import pandas as pd
import numpy as np
import nltk
from nltk.corpus import stopwords
from nltk.stem import WordNetLemmatizer
from nltk.tokenize.casual import casual_tokenize
from vaderSentiment.vaderSentiment import SentimentIntensityAnalyzer
import ast
import requests
import warnings
from tqdm.notebook import tqdm
warnings.filterwarnings('ignore')
tqdm.pandas()

%%html
<style>
.cell-output-ipywidget-background {
    background-color: transparent !important;
}
:root {
    --jp-widgets-color: var(--vscode-editor-foreground);
    --jp-widgets-font-size: var(--vscode-editor-font-size);
}  

df = pd.DataFrame(pd.read_csv('../data/rotten_tomatoes_critic_reviews.csv'))
df.head()

df.dtypes

df.shape

### Data Cleaning

df.isnull().sum()

`critic_name` is not very useful to the project as we do not need to know who wrote the review, just the sentiment. <br>
`review_content`, while important, is not a must to have, and we can fill it in with blank values.

df['review_content'] = df['review_content'].fillna('No review')

df['review_score'].value_counts()

len(df['review_score'].value_counts())

There are 814 unique scores, which is bizzare. <br>
We cannot remove all the rows as it contributes to a significant portion of the data (Roughly 30%). Thus, we cannot use this data.

df2 = df[['rotten_tomatoes_link', 'publisher_name', 'top_critic', 'review_type', 'review_date', 'review_content']]
df2.head()

df2['review_date'] = pd.to_datetime(df2['review_date'])
type(df2['review_date'][0])

nltk.download('punkt', download_dir='./nltk_data', quiet=True)
nltk.download('punkt_tab', download_dir='./nltk_data', quiet=True)
nltk.download('stopwords', download_dir='./nltk_data', quiet=True)
nltk.download('wordnet', download_dir='./nltk_data', quiet=True)
nltk.data.path.append('./nltk_data')

# stop_words = set(stopwords.words('english'))
# stemmer = PorterStemmer()
lemmatizer = WordNetLemmatizer()
stopwords_list = requests.get("https://gist.githubusercontent.com/rg089/35e00abf8941d72d419224cfd5b5925d/raw/12d899b70156fd0041fa9778d657330b024b959c/stopwords.txt").content
stopwords = set(stopwords_list.decode().splitlines())

df2['tokenized_review_content'] = df2['review_content'].progress_apply(lambda x: casual_tokenize(x.lower(), reduce_len=True, strip_handles=False))
df2.head()

df2['alphanum_review_content'] = df2['tokenized_review_content'].progress_apply(lambda x: [word for word in x if word.isalnum()])
df2.head()

df2['stopword_removed_review_content'] = df2['alphanum_review_content'].progress_apply(lambda x: [word for word in x if word not in stopwords])
df2.head()

df2['lemmatized_review_content'] = df2['stopword_removed_review_content'].progress_apply(lambda x: [lemmatizer.lemmatize(word) for word in x])
df2.head()

### VADER

sid_obj = SentimentIntensityAnalyzer()

df2['sentiment_scores'] = df2['review_content'].progress_apply(lambda x: sid_obj.polarity_scores(x))
df2['compound_score'] = df2['sentiment_scores'].progress_apply(lambda x: x['compound'])
df2['sentiment'] = df2['compound_score'].progress_apply(lambda x: 
                                                "Positive" if x > 0 else
                                                ("Negative" if x < 0 else "Neutral"))
df2.head()

df3 = df2[['rotten_tomatoes_link', 'publisher_name', 'top_critic', 'review_type', 'review_date', 'sentiment', 'compound_score']]
df3.to_csv('../data/rotten_tomatoes_critic_reviews_tableau.csv', index=False)

df4 = df2[['top_critic', 'review_type', 'review_date', 'sentiment', 'compound_score', 'lemmatized_review_content']]
df4.to_csv('../data/rotten_tomatoes_critic_reviews_cleaned.csv', index=False)
