==> notebooks/4) Merging Data for Regression Prediction.ipynb <==
## Merging Data

### Import Libraries

import pandas as pd
import numpy as np
import ast
import warnings
from tqdm.notebook import tqdm
warnings.filterwarnings('ignore')
tqdm.pandas()

%%html
<style>
.cell-output-ipywidget-background {
    background-color: transparent !important;
}
:root {
    --jp-widgets-color: var(--vscode-editor-foreground);
    --jp-widgets-font-size: var(--vscode-editor-font-size);
}  

itm_df = pd.DataFrame(pd.read_csv('../data/imdb_tmdb_cleaned_w|o_filtering.csv'))
print(itm_df.shape)
itm_df.head()

rtm_df = pd.DataFrame(pd.read_csv('../data/rotten_tomatoes_movies_cleaned.csv'))
print(rtm_df.shape)
rtm_df.head()

itm_df.columns

### Attribute Processing

itm_df.head()

itm_df['production_companies'] = itm_df['production_companies'].progress_apply(ast.literal_eval)
itm_df['production_countries'] = itm_df['production_countries'].progress_apply(ast.literal_eval)
itm_df['actor_list'] = itm_df['actor_list'].progress_apply(ast.literal_eval)
itm_df['keywords'] = itm_df['keywords'].progress_apply(ast.literal_eval)
itm_df['genre'] = itm_df['genre'].progress_apply(ast.literal_eval)
itm_df['director'] = itm_df['director'].progress_apply(ast.literal_eval)
itm_df.head()

rtm_df.head()

rtm_df.columns

rtm_df = rtm_df[['rotten_tomatoes_link', 'movie_title', 'tomatometer_status', 'tomatometer_rating', 'audience_status', 'audience_rating', 'compound_score', 'sentiment']]
rtm_df.head()

merged_df = pd.merge(itm_df, rtm_df, left_on='title', right_on='movie_title', how='inner')
merged_df = merged_df.drop(columns=['movie_title'])
print(merged_df.shape)

### Exporting Data

itm_df.to_csv('../data/imdb_tmdb_regression.csv', index=False)

merged_df.to_csv('../data/merged_regression.csv', index=False)
