==> notebooks/6) D3 JSON Converter.ipynb <==
## D3.js Dataset to JSON Data Converter

### Import Libraries

import pandas as pd
import numpy as np
import json
import warnings
from tqdm.notebook import tqdm
warnings.filterwarnings('ignore')
tqdm.pandas()

%%html
<style>
.cell-output-ipywidget-background {
    background-color: transparent !important;
}
:root {
    --jp-widgets-color: var(--vscode-editor-foreground);
    --jp-widgets-font-size: var(--vscode-editor-font-size);
}  

df = pd.DataFrame(pd.read_csv('../data/imdb_imdb_tableau.csv'))
df.head()

df2 = pd.DataFrame(pd.read_csv('../data/imdb_tmdb_cleaned.csv'))
df2.head()

df.dtypes

df2.dtypes

df_filtered = df[['title', 'genre']]
df_filtered.head()

df_filtered.rename(columns={'title': 'source', 'genre': 'target'}, inplace=True)
df_filtered.head()

### JSON Conversion Function

def row_to_json(row):
    return row.to_dict()

json_data = df.progress_apply(row_to_json, axis=1).tolist()
print(json_data[0])

json_data2 = df2.progress_apply(row_to_json, axis=1).tolist()
print(json_data2[0])

json_data_filtered = df_filtered.progress_apply(row_to_json, axis=1).tolist()
print(json_data_filtered[0])

### Export Data

with open('../data/imdb_data_exploded.json', 'w') as json_file:
    json.dump(json_data, json_file)

with open('../data/imdb_data_unexploded.json', 'w') as json_file:
    json.dump(json_data2, json_file)

# Export json_data to a JSON file
with open('../data/imdb_data_exploded_filtered.json', 'w') as json_file:
    json.dump(json_data_filtered, json_file)
